bot/requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
